+++ user-service/src/main/java/org/fablewhirl/user/service/UserService.java
@@ -5,13 +5,11 @@ import org.fablewhirl.user.dto.UserCreateEditDto;
 import org.fablewhirl.user.dto.UserReadDto;
 import org.fablewhirl.user.entity.UserEntity;
 import org.fablewhirl.user.event.UserRegisteredEvent;
-import org.fablewhirl.user.mapper.UserCreateEditMapper;
 import org.fablewhirl.user.mapper.UserReadMapper;
 import org.fablewhirl.user.mapper.UserRegisteredEventMapper;
 import org.fablewhirl.user.repository.UserRepository;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
-
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
@@ -23,17 +21,23 @@ import java.util.Optional;
 public class UserService {
 
     private final UserRepository userRepository;
-    private final UserCreateEditMapper userCreateEditMapper;
     private final UserRegisteredEventMapper userRegisteredEventMapper;
     private final UserReadMapper userReadMapper;
     private final PasswordEncoder passwordEncoder;
 
     @Transactional
     public void register(UserRegisteredEvent userDto) {
+        if (existsByUsernameOrEmail(userDto.getUsername(), userDto.getEmail())) {
+            throw new IllegalArgumentException("User with the same username or email already exists");
+        }
+
         UserEntity user = userRegisteredEventMapper.toEntity(userDto);
         user.setPassword(passwordEncoder.encode(userDto.getPassword()));
-
         userRepository.save(user);
+    }
+
+    public boolean existsById(String userId) {
+        return userRepository.existsById(userId);
     }
 
     public boolean existsByUsernameOrEmail(String username, String email) {
@@ -41,9 +45,9 @@ public class UserService {
     }
 
     public UserReadDto getUserById(String id) {
-        return userReadMapper.toDto(
-                userRepository.findById(id)
-                        .orElseThrow(() -> new IllegalArgumentException("User not found")));
+        return userRepository.findById(id)
+                .map(userReadMapper::toDto)
+                .orElseThrow(() -> new IllegalArgumentException("User not found"));
     }
 
     public List<UserEntity> getAll() {
@@ -51,27 +55,33 @@ public class UserService {
     }
 
     @Transactional
-    public UserCreateEditDto updateUser(String id, UserCreateEditDto userRegistrationDto) {
+    public UserReadDto updateUser(String id, UserCreateEditDto userDto) {
         UserEntity userEntity = userRepository.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("User not found"));
 
-        Optional.ofNullable(userRegistrationDto.getUsername())
+        Optional.ofNullable(userDto.getUsername())
                 .ifPresent(userEntity::setUsername);
-        Optional.ofNullable(userRegistrationDto.getPassword())
-                .ifPresent(userEntity::setPassword);
-        Optional.ofNullable(userRegistrationDto.getEmail())
+        Optional.ofNullable(userDto.getPassword())
+                .ifPresent(password -> userEntity.setPassword(passwordEncoder.encode(password)));
+        Optional.ofNullable(userDto.getEmail())
                 .ifPresent(userEntity::setEmail);
-        Optional.ofNullable(userRegistrationDto.getBio())
+        Optional.ofNullable(userDto.getBio())
                 .ifPresent(userEntity::setBio);
+
         userEntity.setUpdatedDate(LocalDateTime.now());
 
-        return userCreateEditMapper.toDto(userRepository.save(userEntity));
+        UserEntity updatedEntity = userRepository.save(userEntity);
+        return userReadMapper.toDto(updatedEntity);
     }
 
     @Transactional
     public void deleteUser(String id) {
+        if (!existsById(id)) {
+            throw new IllegalArgumentException("User not found");
+        }
         userRepository.deleteById(id);
     }
 }
+
